requirements.txt
pdfplumber
requests
python-pptx
